Modesetting: Implement reading EDID data over DisplayPort for DDI A
This implements reading EDID data through the DisplayPort Aux channel
(across which I2C messages are sent).

This works for the internal display on the Switch Alpha 12.  This
fails gracefully on the Intel NUC.

So far, all we do with the EDID data is check the header and print a
log message.

For testing, we implement an emulated device in test_displayport.cc,
which we run the driver against.

MA-150

Change-Id: Ic6a0d9377b925059e5e108fa705191b8428c9cf0