Introduce helper code for accessing fields of MMIO registers
Change the uses of the DdiAuxControl register to use these bitfield
helpers.  This removes some code that manually bit-shifts and masks
register fields.

This will make accessing fields of registers more convenient and
robust.  This will be useful for later modesetting code which will
need to access a lot of registers.

MA-150

Change-Id: I95ee4352c37fdf34a770fbefb786d40a45b739a0