Modesetting: Move entry point call from RenderEngineInit() to Init()
It was an accident that this call was in
MsdIntelDevice::RenderEngineInit() instead of MsdIntelDevice::Init().

In the original pre-commit version, the function call was in Init(),
but it got moved when I rebased on top of the change that introduced
RenderEngineInit() (f717ec1daddd73c252570f1a2c63020e67567810).

MA-150

Change-Id: If6af22c9ca2fd773762fe61f6035b79cd6c82114