Modesetting: Add code for parsing EDID data
The part of the EDID data that I will need in subsequent changes is
timing information -- basically CRT signal timings (which are used
even for modern non-CRT displays).

Add a description of the EDID fields we need, plus tests using some
example EDID data.

Generalise BitfieldRef so that we can easily use it for some of the
8-bit fields in the EDID data.

MA-150

Change-Id: I7dc73c3c7b2b6c8b92d5d02b881fedb28962fd85