Modesetting: Make pipe/plane registers accessible through a PipeRegs class
This is similar to the refactoring I did for the DDI registers.  As
before, this deduplicates the Get() methods of each plane register
class and puts the address calculation in a single place in the
PipeRegs class.

In later changes, I will add some more registers to PipeRegs, so that
it won't only contain plane registers.

MA-150

Change-Id: I943d28bb351123218d24b60bb23336b1b6d3ab02